ui: Remove window around popups
All of the pop-ups generated by MessageBox had non-functional resize
buttons near the title when the demo is running on the RZ/G2M. This was
resolved by creating a MessageBox object rather than using the shorthand
member functions, and adding the Frameless flag. The result is that the
pop-up has no border and resize buttons. But the pop-up cannot be moved,
which may be an issue if resolution causes the pop-up to be outside the
screen.

Signed-off-by: Ayman Chaudhry <ayman.chaudhry.kc@renesas.com>