ui: Place labels inside boxes
When Run is selected the text label appears above the box and can appear
snipped when the item is located near the top of the frame. By making
the label appear inside the box, the label is restricted to the
parameters of the box and does not appear out of the frame.

Signed-off-by: Ayman Chaudhry <ayman.chaudhry.kc@renesas.com>