ui: Clear on webcam disconnect
If the webcam disconnect button is pressed the application will continue
to display the last known frame, boxes and table. Checking if the webcam
disconnect button has been pressed when drawing boxes allows the
application to clear the display. The webcam buttons can be enabled
again by pressing the webcam reset button.
Signed-off-by: Ayman Chaudhry <ayman.chaudhry.kc@renesas.com>