Add support for live video preview
Video feed is shown at start. It gets stopped when the "Process Basket"
button is pressed and restarted when the "Next Basket" button is pressed.

At the moment there is no limit on the speed of the video loop, so one of
the CPU cores will be maxed out.

Thread setup code adapted from
https://stackoverflow.com/questions/14051276/how-to-implement-frequent-start-stop-of-a-thread-qthread

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>