webcam: Correct default camera logic
In the instance that the /dev/v4l/by-id directory does not exist
the camera dev path provided by the user at terminal will not be
used by the application because of the logical AND in use. We
should correct this logic to separate the user option from the
filesystem.

Signed-off-by: Gareth Williams <gareth.williams.jx@renesas.com>